Move sparklite presskit files into same folder as other presskits
Went back and forth on whether these files belong with the presskit
or in the assets, since the Sparklite page might reference them. In the
end I think it's hard to get this 100% but want to make sure the presskit
files are easy to track down and maintain so keeping them together made
sense.